feat/refactor: 리뷰 조회시 페이지네이션 구현 및 코드 수정

+++ src/main/java/com/culture/BAEUNDAY/domain/user/DTO/response/RegisterResponseDTO.java
@@ -8,7 +8,7 @@ public record RegisterResponseDTO(
         String name,
         String username,
         String profileImg,
-        Double manner
+        Integer manner
 ) {
